refactor: update image-card
