+++ src/app/my/_components/store-counter-card.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import ImageCard from "@/components/supplier/image-card";
+import ImageCardPrimitive from "@/components/image-card/image-card-primitive";
 import useStore from "@/hooks/use-store";
 import type { SelectStoreDish } from "@/lib/type";
 
@@ -11,7 +11,7 @@ export function StoreCounterCard({
 }) {
   const { updateStoreDish } = useStore();
   return (
-    <ImageCard
+    <ImageCardPrimitive
       href={`/reservation/${storeDish.id}`}
       image={storeDish.image}
       className="border-none shadow-none"
@@ -27,6 +27,6 @@ export function StoreCounterCard({
       <div className="mt-1 line-clamp-3 w-full text-xs text-muted-foreground">
         {storeDish.description}
       </div>
-    </ImageCard>
+    </ImageCardPrimitive>
   );
 }
